kernel: remove vm subsystem and rollback limine revision
Taking a quick detour here for the kernel roadmap, I
noticed it would be much easier to have limine map
the early memory structures for us, so we can have the
scheduler running before the memory manager. Therefore,
I switched the limine revision back to 2 (which maps
all memory into the higher half) and removed the initial
memory subsystem code.